refactor(core): mocks cleanup
