+++ packages/core/lib/mediator/mediator.spec.ts
@@ -3,8 +3,9 @@ import { of, Subject, toArray } from 'rxjs';
 import { MessageTypes } from '../messages';
 import { ProvidersManagerMock } from '../providers-manager/providers-manager.mocks';
 import { ExtensionsManagerMock } from '../extensions/extensions-manager.mocks';
+import { TestRequest } from '../messages/request/messages.mocks';
 import { ExecutorMock } from '../executor/executor.mocks';
-import { TestEvent, TestRequest } from '../messages/messages.mocks';
+import { TestEvent } from '../messages/event/events.mocks';
 import { IExecutor } from '../executor/ports/executor.port';
 import { IMediator } from './ports/mediator.port';
 import { Mediator } from './mediator';
