+++ packages/core/lib/messages/event/decorators/global-event-handler.decorator.spec.ts
@@ -1,8 +1,8 @@
 import 'reflect-metadata';
 import { SCOPE_OPTIONS_METADATA } from '../../constants';
 import { ScopeOptions } from '../../interfaces';
-import { TestEventHandler } from '../../messages.mocks';
 import { GLOBAL_EVENT_HANDLER_METADATA } from '../constants';
+import { TestEventHandler } from '../events.mocks';
 import { GlobalEventHandler } from './global-event-handler.decorator';
 
 describe('GlobalEventHandler', () => {
