+++ packages/core/lib/messages/event/events.mocks.ts
@@ -1,0 +1,7 @@
+import { IEventHandler } from './interfaces';
+
+export class TestEvent {}
+
+export class TestEventHandler implements IEventHandler<TestEvent> {
+  handle = jest.fn(async (): Promise<any> => undefined);
+}
