+++ packages/core/lib/messages/request/decorators/request-pipeline.decorator.spec.ts
@@ -1,8 +1,8 @@
 import 'reflect-metadata';
 import { SCOPE_OPTIONS_METADATA } from '../../constants';
 import { ScopeOptions } from '../../interfaces';
-import { TestRequest, TestRequestPipeline } from '../../messages.mocks';
 import { REQUEST_PIPELINE_METADATA } from '../constants';
+import { TestRequest, TestRequestPipeline } from '../messages.mocks';
 import { RequestPipeline } from './request-pipeline.decorator';
 
 class SomeRequest {}
